requirements.txt
requests
python-dotenv
xgboost
pandas
numpy
scikit-learn
streamlit
folium
polyline
streamlit-folium
plotly
fastapi
uvicorn
websockets
